Transaction saving logic was implemented
Local update balance logic was implemented
forms initialization was improved

+++ Practice5_BabyBlockchain/DuckCoin/DuckCoin.Wallet/DomainModels/Account.cs
@@ -18,7 +18,7 @@ namespace DuckCoin.Wallet.DomainModels
 
         public string PrivateKey { get; set; }
 
-        public int Balance { get; set; }
+        public double Balance { get; set; }
 
         public string PasswordHash { get; set; }
